Update app
- Structure for running locally and deploying to production
- Secure connection string
- Build out README
- Update language on index
