+++ .gitignore
@@ -26,4 +26,10 @@ ENV/
 env.bak/
 venv.bak/
 
-/flask/cockroach*
+cockroach*
+certs/
+my-safe-directory/
+build/
+dist/
+*.egg-info/
+__pycache__
