+++ README.md
@@ -7,7 +7,7 @@ the code used for various parts of the book:
 starting points for the poems IF and ELSE and to generate the poem
 2019-04-06-11:05:32 UTC (unmodified and named for its generation time), based on
 the remnants of the process of writing the poem IN.
-- The `flask` directory contains the Flask web app version, which is OUT, the
+- The `flask` directory contains the Flask web app version, which is the
 the last poem of the book.
 
 See the `README.md` files in those directories for more details.
