+++ flask/MANIFEST.in
@@ -1,0 +1,5 @@
+include rotations/schema.sql
+include rotations/source.txt
+graft rotations/static
+graft rotations/templates
+global-exclude *.pyc
