+++ flask/rotations/templates/poem/index.html
@@ -1,10 +1,10 @@
 <nav>
-  <h1 class="index">OUT</h1>
+  <h1 class="index">SEVERAL ROTATIONS</h1>
 </nav>
 {% extends 'base.html' %}
 {% block content %}
   </br>
-  <p>This is the last poem of the book SEVERAL ROTATIONS. If you've read the first poems of the book, what you generate here can be the end for you, or not.</p>
+  <p>This is the last poem of the book SEVERAL ROTATIONS. If you've read the first poems of the book, what you generate here can be the end of the book for you, or not.</p>
   <p>Everything you or anyone else generates here is stored in a database and is a cumulative and continuous end to the book as well. Eventually, some or all of it might show up here.</p>
   <p>The code for this site is at <a href="https://github.com/jseldess/several-rotations/tree/master/flask" target="_blank">https://github.com/jseldess/several-rotations</a>.</p>
   </br>
