+++ flask/setup.py
@@ -1,0 +1,13 @@
+from setuptools import find_packages, setup
+
+setup(
+    name='rotations',
+    version='1.0.0',
+    packages=find_packages(),
+    include_package_data=True,
+    zip_safe=False,
+    install_requires=[
+        'flask',
+        'psycopg2-binary',
+    ],
+)
